SQUASH: ordination
