+++ q2_types/ordination/__init__.py
@@ -1,0 +1,19 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2016--, QIIME development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+
+import importlib
+
+from ._format import OrdinationFormat, OrdinationDirectoryFormat
+from ._type import PCoAResults
+
+
+__all__ = ['OrdinationFormat', 'OrdinationDirectoryFormat',
+           'PCoAResults']
+
+
+importlib.import_module('q2_types.ordination._transformer')
